Move token preperation logic to a concern
The methods to authorize and set the package/project
from a token is also required in the webui controller.
Therefore moving it to a concern in order to reuse them.
