+++ src/api/app/controllers/trigger_controller.rb
@@ -1,4 +1,6 @@
 class TriggerController < ApplicationController
+  include Triggerable
+
   ALLOWED_GITLAB_EVENTS = ['Push Hook', 'Tag Push Hook', 'Merge Request Hook'].freeze
 
   include Pundit
@@ -19,10 +21,12 @@ class TriggerController < ApplicationController
   before_action :set_object_to_authorize
   # set_multibuild_flavor needs to run after the set_object_to_authorize callback
   append_before_action :set_multibuild_flavor
+
   include Trigger::Errors
 
   def create
     authorize @token, :trigger?
+
     @token.user.run_as do
       opts = { project: @project, package: @package, repository: params[:repository], arch: params[:arch] }
       opts[:multibuild_flavor] = @multibuild_container if @multibuild_container.present?
@@ -59,47 +63,5 @@ class TriggerController < ApplicationController
 
   def pundit_user
     @token.user
-  end
-
-  def set_project
-    # By default we operate on the package association
-    @project = @token.package.try(:project)
-    # If the token has no package, let's find one from the parameters
-    @project ||= Project.get_by_name(params[:project])
-    # Remote projects are read-only, can't trigger something for them.
-    # See https://github.com/openSUSE/open-build-service/wiki/Links#project-links
-    raise Project::Errors::UnknownObjectError, "Sorry, triggering tokens for remote project \"#{params[:project]}\" is not possible." unless @project.is_a?(Project)
-  end
-
-  def set_package
-    # By default we operate on the package association
-    @package = @token.package
-    # If the token has no package, let's find one from the parameters
-    @package ||= Package.get_by_project_and_name(@project,
-                                                 params[:package],
-                                                 @token.package_find_options)
-    return unless @project.links_to_remote?
-
-    # The token has no package, we did not find a package in the database but the project has a link to remote.
-    # See https://github.com/openSUSE/open-build-service/wiki/Links#project-links
-    # In this case, we will try to trigger with the user input, no matter what it is
-    @package ||= params[:package]
-
-    # TODO: This should not happen right? But who knows...
-    raise ActiveRecord::RecordNotFound unless @package
-  end
-
-  def set_object_to_authorize
-    @token.object_to_authorize = package_from_project_link? ? @project : @package
-  end
-
-  def set_multibuild_flavor
-    # Do NOT use @package.multibuild_flavor? here because the flavor need to be checked for the right source revision
-    @multibuild_container = params[:package].gsub(/.*:/, '') if params[:package].present? && params[:package].include?(':')
-  end
-
-  def package_from_project_link?
-    # a remote package is always included via project link
-    !(@package.is_a?(Package) && @package.project == @project)
   end
 end
